fix: don't check host state for errors that cause shutdown
https://gerrit.openbmc-project.xyz/c/openbmc/host-error-monitor/+/31533

Don't check host state for errors that cause shutdown

Some falling-edge GPIOs go low when the system powers off
causing false errors to be detected, so host state must be
checked when handling.

However, there are some falling-egde SGPIOs that the hardware
is able to keep high when the system powers off, so the host
state doesn't need to be checked.  On top of that, if the error
itself causes the system to shut down, the event could be
missed by checking the host state.

This commit removes the host state check for the signals that
stay high when the system powers off and may cause a shutdown
when asserted.

Signed-off-by: George Hung <george.hung@quantatw.com>